+++ ayunis-core-backend/src/iam/invites/application/use-cases/create-bulk-invites/create-bulk-invites.use-case.ts
@@ -19,6 +19,7 @@ import { ApplicationError } from 'src/common/errors/base.error';
 import { UsersRepository } from 'src/iam/users/application/ports/users.repository';
 import { SubscriptionNotFoundError } from 'src/iam/subscriptions/application/subscription.errors';
 import { UserRole } from 'src/iam/users/domain/value-objects/role.object';
+import { getInviteExpiresAt } from '../../services/invite-expiration.util';
 
 interface BulkInviteResult {
   email: string;
@@ -283,7 +284,7 @@ export class CreateBulkInvitesUseCase {
     );
 
     // Create all invite entities upfront
-    const inviteExpiresAt = this.getInviteExpiresAt(validDuration);
+    const inviteExpiresAt = getInviteExpiresAt(validDuration);
     const invites: Invite[] = command.invites.map(
       (inviteData) =>
         new Invite({
@@ -420,39 +421,5 @@ export class CreateBulkInvitesUseCase {
     }
 
     return results;
-  }
-
-  private getInviteExpiresAt(inviteExpiresIn: string): Date {
-    // Parse duration like "7d", "24h", "60m", "3600s"
-    const match = inviteExpiresIn.match(/^(\d+)([dhms])$/);
-
-    if (!match) {
-      throw new Error(
-        `Invalid invite expires in format: ${inviteExpiresIn}. Expected format: number + d/h/m/s (e.g., "7d", "24h")`,
-      );
-    }
-
-    const [, amountStr, unit] = match;
-    const amount = parseInt(amountStr, 10);
-
-    let multiplier: number;
-    switch (unit) {
-      case 'd':
-        multiplier = 24 * 60 * 60 * 1000; // days to milliseconds
-        break;
-      case 'h':
-        multiplier = 60 * 60 * 1000; // hours to milliseconds
-        break;
-      case 'm':
-        multiplier = 60 * 1000; // minutes to milliseconds
-        break;
-      case 's':
-        multiplier = 1000; // seconds to milliseconds
-        break;
-      default:
-        throw new Error(`Unsupported time unit: ${unit}`);
-    }
-
-    return new Date(Date.now() + amount * multiplier);
   }
 }
