+++ ayunis-core-backend/src/iam/invites/application/use-cases/create-invite/create-invite.use-case.ts
@@ -20,6 +20,7 @@ import { FindUserByEmailUseCase } from 'src/iam/users/application/use-cases/find
 import { FindUserByEmailQuery } from 'src/iam/users/application/use-cases/find-user-by-email/find-user-by-email.query';
 import { UserEmailProviderBlacklistedError } from 'src/iam/users/application/users.errors';
 import { SubscriptionNotFoundError } from 'src/iam/subscriptions/application/subscription.errors';
+import { getInviteExpiresAt } from '../../services/invite-expiration.util';
 
 interface CreateInviteResult {
   token: string;
@@ -111,7 +112,7 @@ export class CreateInviteUseCase {
         'auth.jwt.inviteExpiresIn',
         '7d',
       );
-      const inviteExpiresAt = this.getInviteExpiresAt(validDuration);
+      const inviteExpiresAt = getInviteExpiresAt(validDuration);
 
       const invite = new Invite({
         email: command.email,
@@ -147,39 +148,5 @@ export class CreateInviteUseCase {
       });
       throw new UnexpectedInviteError(error as Error);
     }
-  }
-
-  private getInviteExpiresAt(inviteExpiresIn: string): Date {
-    // Parse duration like "7d", "24h", "60m", "3600s"
-    const match = inviteExpiresIn.match(/^(\d+)([dhms])$/);
-
-    if (!match) {
-      throw new Error(
-        `Invalid invite expires in format: ${inviteExpiresIn}. Expected format: number + d/h/m/s (e.g., "7d", "24h")`,
-      );
-    }
-
-    const [, amountStr, unit] = match;
-    const amount = parseInt(amountStr, 10);
-
-    let multiplier: number;
-    switch (unit) {
-      case 'd':
-        multiplier = 24 * 60 * 60 * 1000; // days to milliseconds
-        break;
-      case 'h':
-        multiplier = 60 * 60 * 1000; // hours to milliseconds
-        break;
-      case 'm':
-        multiplier = 60 * 1000; // minutes to milliseconds
-        break;
-      case 's':
-        multiplier = 1000; // seconds to milliseconds
-        break;
-      default:
-        throw new Error(`Unsupported time unit: ${unit}`);
-    }
-
-    return new Date(Date.now() + amount * multiplier);
   }
 }
